# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/record": {
      "post": {
        "operationId": "RecordController_createRecord",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateRecordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "record"
        ]
      },
      "get": {
        "operationId": "RecordController_getRecordOfAllHabit",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "record"
        ]
      }
    },
    "/record/grass": {
      "get": {
        "operationId": "RecordController_getGrass",
        "parameters": [
          {
            "name": "months",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "습관별 최근 N개월 기록 (잔디)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GrassItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "record"
        ]
      }
    },
    "/record/{id}": {
      "get": {
        "operationId": "RecordController_getRecordOfOneHabit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "record"
        ]
      },
      "delete": {
        "operationId": "RecordController_removeRecord",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "record"
        ]
      }
    }
  },
  "info": {
    "title": "Happit API",
    "description": "Happit API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrUpdateRecordDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Record date (YYYY-MM-DD)",
            "example": "2025-02-15"
          },
          "habitId": {
            "type": "number",
            "description": "Habit id"
          },
          "state": {
            "type": "string",
            "description": "Completion state",
            "enum": [
              "done",
              "notDone",
              "skipped"
            ]
          }
        },
        "required": [
          "date",
          "habitId",
          "state"
        ]
      },
      "GrassRecordItemDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2025-02-15"
          },
          "state": {
            "type": "string",
            "enum": [
              "done",
              "notDone",
              "skipped"
            ]
          }
        },
        "required": [
          "date",
          "state"
        ]
      },
      "GrassItemDto": {
        "type": "object",
        "properties": {
          "habitId": {
            "type": "number"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrassRecordItemDto"
            }
          }
        },
        "required": [
          "habitId",
          "records"
        ]
      }
    }
  }
}